+++ src/Utils.ts
@@ -1,3 +1,5 @@
+import * as fs from 'fs';
+import * as path from 'path';
 import * as url from 'url';
 
 export default class Utils {
@@ -595,6 +597,36 @@ export default class Utils {
         filtered[key] = obj[key];
         return filtered;
       }, {});
+  }
+
+  public static readdirR(dir: string): string | string[] {
+    return fs.statSync(dir).isDirectory()
+      ? Array.prototype.concat(...fs.readdirSync(dir).map(f => Utils.readdirR(path.join(dir, f))))
+      : dir;
+  }
+
+  // from: https://stackoverflow.com/a/22185855
+  public static copyRecursiveSync(src: string, dest: string, replaceTokens?: (s: string) => string): void {
+    const exists: boolean = fs.existsSync(src);
+    const stats: false | fs.Stats = exists && fs.statSync(src);
+    const isDirectory: boolean = exists && (stats as fs.Stats).isDirectory();
+
+    if (replaceTokens) {
+      dest = replaceTokens(dest);
+    }
+
+    if (isDirectory) {
+      if (!fs.existsSync(dest)) {
+        fs.mkdirSync(dest);
+      }
+      fs.readdirSync(src).forEach(function (childItemName) {
+        Utils.copyRecursiveSync(path.join(src, childItemName),
+          path.join(dest, childItemName), replaceTokens);
+      });
+    }
+    else {
+      fs.copyFileSync(src, dest);
+    }
   }
 
   public static parseCsvToJson(s: string): any {