+++ docs/docs/about/release-notes.md
@@ -11,6 +11,7 @@
 
 **CLI:**
 
+- [cli config reset](../cmd/cli/config/config-reset.md) - resets the specified CLI configuration option to its default value [#2344](https://github.com/pnp/cli-microsoft365/issues/2344)
 - [cli issue](../cmd/cli/cli-issue.md) - returns, or opens a URL that takes the user to the right place in the CLI GitHub repo to create a new issue reporting bug, feedback, ideas, etc. [#1490](https://github.com/pnp/cli-microsoft365/issues/1490)
 
 **Microsoft 365:**
@@ -18,6 +19,10 @@
 - [tenant report office365activationcounts](../cmd/tenant/report/report-office365activationcounts.md) - gets the count of Microsoft 365 activations on desktops and devices [#2257](https://github.com/pnp/cli-microsoft365/issues/2257)
 - [tenant report office365activationsusercounts](../cmd/tenant/report/report-office365activationsusercounts.md) - gets the count of users that are enabled and those that have activated the Office subscription on desktop or devices or shared computers [#2258](https://github.com/pnp/cli-microsoft365/issues/2258)
 - [tenant report office365activationsuserdetail](../cmd/tenant/report/report-office365activationsuserdetail.md) - Get details about users who have activated Microsoft 365 [#2256](https://github.com/pnp/cli-microsoft365/issues/2256)
+
+**SharePoint Framework:**
+
+- [spfx package generate](../cmd/spfx/package/package-generate.md) - generates SharePoint Framework solution package with a no-framework web part rendering the specified HTML snippet [#2367](https://github.com/pnp/cli-microsoft365/issues/2367)
 
 ### Changes
 
@@ -34,6 +39,7 @@
 - extended 'spo file list' with the `recursive` option [#2376](https://github.com/pnp/cli-microsoft365/issues/2376)
 - added the Get-SPOUser CLI equivalents to the docs [#2485](https://github.com/pnp/cli-microsoft365/issues/2485)
 - removed the deprecated Get-TeamHelp cmdlet from the docs [#2486](https://github.com/pnp/cli-microsoft365/issues/2486)
+- added the Remove-PowerApp CLI equivalent to the docs [#2491](https://github.com/pnp/cli-microsoft365/issues/2491)
 
 ## [v3.10.0](https://github.com/pnp/cli-microsoft365/releases/tag/v3.10.0)